finished ui design and fixes, pushed prod table to database
