+++ app/lib/guests.js
@@ -1,5 +1,4 @@
-const guest_table = 
-[
+const guest_table = [
 	{
 		guest_id: "aswed0",
 		name: "Avani Pai",
@@ -11,19 +10,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -39,19 +38,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -67,19 +66,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -95,19 +94,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -123,19 +122,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -151,19 +150,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -179,19 +178,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -207,19 +206,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -235,19 +234,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -263,19 +262,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -291,19 +290,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -319,19 +318,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -347,19 +346,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -375,19 +374,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -403,19 +402,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -431,19 +430,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -459,19 +458,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -487,19 +486,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -515,19 +514,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -543,19 +542,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -571,19 +570,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
@@ -599,25 +598,53 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
 		guest_id: "aswed22",
+		name: "Kiyomi Toda",
+		email: "kiyo410810sid@gmail.com",
+		password: "fionashome301",
+		ReceptionOnly: false,
+		SangeetReceptionOnly: false,
+		SanMuhRec: false,
+		AllEvents: false,
+		AllUSJapan: true,
+		isAttendingMehendi: "no",
+		mehendiPartySize: "0",
+		isAttendingHaldi: "no",
+		haldiPartySize: "0",
+		isAttendingSangeet: "no",
+		sangeetPartySize: "0",
+		isAttendingMuhurtham: "no",
+		muhurthamPartySize: "0",
+		isAttendingReception: "no",
+		receptionPartySize: "0",
+		isAttendingShinzenshiki: "no",
+		shinzenshikiPartySize: "0",
+		isAttendingHiroen: "no",
+		hiroenPartySize: "0",
+		lastRSVPUpdateTime: "2024-05-26 10:00",
+		partyMembers: "avani",
+		allergies: "none"
+	},
+	{
+		guest_id: "aswed23",
 		name: "Dilip & Anusha",
 		email: "dmallya93@gmail.com",
 		password: "fionashome301",
@@ -627,25 +654,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed23",
+		guest_id: "aswed24",
 		name: "Ganesh Pai",
 		email: "ganeshrpai@gmail.com",
 		password: "fionashome301",
@@ -655,25 +682,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed24",
+		guest_id: "aswed25",
 		name: "Latha Pai",
 		email: "lpai265@gmail.com",
 		password: "fionashome301",
@@ -683,25 +710,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed25",
+		guest_id: "aswed26",
 		name: "Oorvi Pai",
 		email: "oorvi.pai99@gmail.com",
 		password: "fionashome301",
@@ -711,25 +738,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed26",
+		guest_id: "aswed27",
 		name: "Aparna Pai",
 		email: "pai.pyro@gmail.com",
 		password: "fionashome301",
@@ -739,25 +766,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed27",
+		guest_id: "aswed28",
 		name: "Harish Pai",
 		email: "h_pai@yahoo.com",
 		password: "fionashome301",
@@ -767,25 +794,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed28",
+		guest_id: "aswed29",
 		name: "Girish Pai",
 		email: "girpai@gmail.com",
 		password: "fionashome301",
@@ -795,25 +822,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed29",
+		guest_id: "aswed30",
 		name: "Lotika Pai",
 		email: "lotika.pai@gmail.com",
 		password: "fionashome301",
@@ -823,25 +850,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed30",
+		guest_id: "aswed31",
 		name: "Indraneel Pai",
 		email: "Neel227@gmail.com",
 		password: "fionashome301",
@@ -851,25 +878,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed31",
+		guest_id: "aswed32",
 		name: "Nachiketa Pai",
 		email: "nachipai@gmail.com",
 		password: "fionashome301",
@@ -879,25 +906,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed32",
+		guest_id: "aswed33",
 		name: "Alexis Broedlow",
 		email: "lbroedlow@gmail.com",
 		password: "fionashome301",
@@ -907,25 +934,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed33",
+		guest_id: "aswed34",
 		name: "Mahiro O'Haru",
 		email: "mahirooharu@gmail.com",
 		password: "fionashome301",
@@ -935,25 +962,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed34",
+		guest_id: "aswed35",
 		name: "Alyssa Liu",
 		email: "ljhannn@gmail.com",
 		password: "fionashome301",
@@ -963,25 +990,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed35",
+		guest_id: "aswed36",
 		name: "Graham Repp",
 		email: "grahamjrepp@gmail.com",
 		password: "fionashome301",
@@ -991,25 +1018,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed36",
+		guest_id: "aswed37",
 		name: "Santhosh Kempanna",
 		email: "mkspavash@gmail.com",
 		password: "fionashome301",
@@ -1019,25 +1046,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed37",
+		guest_id: "aswed38",
 		name: "Ashwin Siddarth",
 		email: "ashwinsid6@gmail.com",
 		password: "fionashome301",
@@ -1047,25 +1074,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed38",
+		guest_id: "aswed39",
 		name: "Divakar Mallya",
 		email: "divakarmallya2000@gmail.com",
 		password: "fionashome301",
@@ -1075,25 +1102,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed39",
+		guest_id: "aswed40",
 		name: "Michelle Mohamed",
 		email: "michellemohamed3@gmail.com",
 		password: "fionashome301",
@@ -1103,25 +1130,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed40",
+		guest_id: "aswed41",
 		name: "Ayusha Ayalur",
 		email: "ayusha.ayalur@gmail.com",
 		password: "fionashome301",
@@ -1131,25 +1158,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: true,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed41",
+		guest_id: "aswed42",
 		name: "Ganesh Pai",
 		email: "ganeshpai@ektelesi.com",
 		password: "fionashome301",
@@ -1159,25 +1186,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed42",
+		guest_id: "aswed43",
 		name: "Ashwini & Abhishek Harish",
 		email: "Abhishek1703@gmail.com",
 		password: "fionashome301",
@@ -1187,25 +1214,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed43",
+		guest_id: "aswed44",
 		name: "Yamuna Nagaraj & Nataraj Lakshman",
 		email: "Yamuna.nagaraj@gmail.com",
 		password: "fionashome301",
@@ -1215,25 +1242,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed44",
+		guest_id: "aswed45",
 		name: "Anika Riar & Sujit Acharya",
 		email: "Anika.riar@gmail.com",
 		password: "fionashome301",
@@ -1243,25 +1270,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed45",
+		guest_id: "aswed46",
 		name: "Chitra Beliiiappa & Satish",
 		email: "chitsbelliappa@gmail.com",
 		password: "fionashome301",
@@ -1271,25 +1298,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed46",
+		guest_id: "aswed47",
 		name: "Kushi",
 		email: "gnanashree.30@gmail.com",
 		password: "fionashome301",
@@ -1299,25 +1326,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed47",
+		guest_id: "aswed48",
 		name: "Padmini & Shailesh Dighe",
 		email: "pchittip@yahoo.com",
 		password: "fionashome301",
@@ -1327,25 +1354,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed48",
+		guest_id: "aswed49",
 		name: "Preeti & Ganesh Kini",
 		email: "preetikini@yahoo.com",
 		password: "fionashome301",
@@ -1355,25 +1382,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed49",
+		guest_id: "aswed50",
 		name: "Saritha & Mayur",
 		email: "Itssaritha@gmail.com",
 		password: "fionashome301",
@@ -1383,25 +1410,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed50",
+		guest_id: "aswed51",
 		name: "Satin & Sonia Manik",
 		email: "Smanik1@hotmail.com",
 		password: "fionashome301",
@@ -1411,25 +1438,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed51",
+		guest_id: "aswed52",
 		name: "Sunitha & Suresh Rajan",
 		email: "Sunitharajan@yahoo.com",
 		password: "fionashome301",
@@ -1439,25 +1466,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed52",
+		guest_id: "aswed53",
 		name: "Seema & Satin Sanghi",
 		email: "satinseema@yahoo.com",
 		password: "fionashome301",
@@ -1467,25 +1494,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed53",
+		guest_id: "aswed54",
 		name: "Adithi & Murlidhar Shenoy",
 		email: "Murlidhar.shenoy@gmail.com",
 		password: "fionashome301",
@@ -1495,25 +1522,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed54",
+		guest_id: "aswed55",
 		name: "Uma Saraf",
 		email: "saraffuma@gmail.com",
 		password: "fionashome301",
@@ -1523,25 +1550,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed55",
+		guest_id: "aswed56",
 		name: "Kiran & Suma Saraf",
 		email: "limcaky@gmail.com",
 		password: "fionashome301",
@@ -1551,25 +1578,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed56",
+		guest_id: "aswed57",
 		name: "Sumitha Bhandarkar",
 		email: "sumithabhandarkar@yahoo.com",
 		password: "fionashome301",
@@ -1579,25 +1606,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed57",
+		guest_id: "aswed58",
 		name: "Vikas & Aparna Guragunti",
 		email: "Vikas95014@gmail.com",
 		password: "fionashome301",
@@ -1607,25 +1634,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed58",
+		guest_id: "aswed59",
 		name: "Chinki",
 		email: "vnimmane@gmail.com",
 		password: "fionashome301",
@@ -1635,25 +1662,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed59",
+		guest_id: "aswed60",
 		name: "Ashika",
 		email: "ashika1808@icloud.com",
 		password: "fionashome301",
@@ -1663,25 +1690,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed60",
+		guest_id: "aswed61",
 		name: "Deepu",
 		email: "deepika.x.mathur@gmail.com",
 		password: "fionashome301",
@@ -1691,25 +1718,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed61",
+		guest_id: "aswed62",
 		name: "Avanthi Kamat",
 		email: "avanthikamat@gmail.com",
 		password: "fionashome301",
@@ -1719,25 +1746,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed62",
+		guest_id: "aswed63",
 		name: "Jairamdas Pai",
 		email: "mjairamdaspai@gmail.com",
 		password: "fionashome301",
@@ -1747,25 +1774,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed63",
+		guest_id: "aswed64",
 		name: "Akshatha",
 		email: "akshathabhat@outlook.com",
 		password: "fionashome301",
@@ -1775,25 +1802,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed64",
+		guest_id: "aswed65",
 		name: "Bhavya",
 		email: "bhavya26@gmail.com",
 		password: "fionashome301",
@@ -1803,25 +1830,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed65",
+		guest_id: "aswed66",
 		name: "Ashwini",
 		email: "ashbhat18@gmail.com",
 		password: "fionashome301",
@@ -1831,25 +1858,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed66",
+		guest_id: "aswed67",
 		name: "Mahesh Prabhu",
 		email: "prabhu.mahesh@gmail.com",
 		password: "fionashome301",
@@ -1859,25 +1886,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed67",
+		guest_id: "aswed68",
 		name: "Deepa Bhat",
 		email: "deepabhat207@gmail.com",
 		password: "fionashome301",
@@ -1887,25 +1914,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed68",
+		guest_id: "aswed69",
 		name: "Nivideta & Vinayak",
 		email: "nivi.prabhu.np@gmail.com",
 		password: "fionashome301",
@@ -1915,25 +1942,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed69",
+		guest_id: "aswed70",
 		name: "Carmen",
 		email: "carmen_cuevas@hotmail.com",
 		password: "fionashome301",
@@ -1943,25 +1970,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed70",
+		guest_id: "aswed71",
 		name: "Andres & Marsha",
 		email: "andresconejeroconesa@gmail.com",
 		password: "fionashome301",
@@ -1971,25 +1998,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed71",
+		guest_id: "aswed72",
 		name: "Penny & Rick Ose",
 		email: "rpmcose@yahoo.com",
 		password: "fionashome301",
@@ -1999,25 +2026,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed72",
+		guest_id: "aswed73",
 		name: "Abhi & Nitesh Arora",
 		email: "abhiaro@gmail.com",
 		password: "fionashome301",
@@ -2027,25 +2054,25 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed73",
+		guest_id: "aswed74",
 		name: "Dave Mitchell",
 		email: "davemitchellchocolate@gmail.com",
 		password: "fionashome301",
@@ -2055,25 +2082,81 @@ const guest_table =
 		AllEvents: true,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed74",
+		guest_id: "aswed75",
+		name: "Avani Shah",
+		email: "pritisweetweds@gmail.com",
+		password: "fionashome301",
+		ReceptionOnly: false,
+		SangeetReceptionOnly: false,
+		SanMuhRec: false,
+		AllEvents: true,
+		AllUSJapan: false,
+		isAttendingMehendi: "no",
+		mehendiPartySize: "0",
+		isAttendingHaldi: "no",
+		haldiPartySize: "0",
+		isAttendingSangeet: "no",
+		sangeetPartySize: "0",
+		isAttendingMuhurtham: "no",
+		muhurthamPartySize: "0",
+		isAttendingReception: "no",
+		receptionPartySize: "0",
+		isAttendingShinzenshiki: "no",
+		shinzenshikiPartySize: "0",
+		isAttendingHiroen: "no",
+		hiroenPartySize: "0",
+		lastRSVPUpdateTime: "2024-05-26 10:00",
+		partyMembers: "avani",
+		allergies: "none"
+	},
+	{
+		guest_id: "aswed76",
+		name: "For Vendors",
+		email: "avaniandshusuke2024@gmail.com",
+		password: "fionashome301",
+		ReceptionOnly: false,
+		SangeetReceptionOnly: false,
+		SanMuhRec: false,
+		AllEvents: true,
+		AllUSJapan: false,
+		isAttendingMehendi: "no",
+		mehendiPartySize: "0",
+		isAttendingHaldi: "no",
+		haldiPartySize: "0",
+		isAttendingSangeet: "no",
+		sangeetPartySize: "0",
+		isAttendingMuhurtham: "no",
+		muhurthamPartySize: "0",
+		isAttendingReception: "no",
+		receptionPartySize: "0",
+		isAttendingShinzenshiki: "no",
+		shinzenshikiPartySize: "0",
+		isAttendingHiroen: "no",
+		hiroenPartySize: "0",
+		lastRSVPUpdateTime: "2024-05-26 10:00",
+		partyMembers: "avani",
+		allergies: "none"
+	},
+	{
+		guest_id: "aswed77",
 		name: "Ganesh Pai",
 		email: "pairganesh@yahoo.com",
 		password: "fionashome301",
@@ -2083,25 +2166,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed75",
+		guest_id: "aswed78",
 		name: "Sunita",
 		email: "sunita.kavthekar@gmail.com",
 		password: "fionashome301",
@@ -2111,25 +2194,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed76",
+		guest_id: "aswed79",
 		name: "Abishek",
 		email: "abishekvenkit@gmail.com",
 		password: "fionashome301",
@@ -2139,25 +2222,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed77",
+		guest_id: "aswed80",
 		name: "Abby Garske",
 		email: "abbygarske5@gmail.com",
 		password: "fionashome301",
@@ -2167,25 +2250,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed78",
+		guest_id: "aswed81",
 		name: "Emma Riedl",
 		email: "emma.b.riedl@gmail.com",
 		password: "fionashome301",
@@ -2195,25 +2278,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed79",
+		guest_id: "aswed82",
 		name: "Juan Antonio Montesinos",
 		email: "jamdm64@gmail.com",
 		password: "fionashome301",
@@ -2223,25 +2306,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed80",
+		guest_id: "aswed83",
 		name: "Jennifer & Matt Kroll",
 		email: "mattkroll@outlook.com",
 		password: "fionashome301",
@@ -2251,25 +2334,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed81",
+		guest_id: "aswed84",
 		name: "Scott Novak",
 		email: "sjn216@icloud.com",
 		password: "fionashome301",
@@ -2279,25 +2362,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed82",
+		guest_id: "aswed85",
 		name: "Katie & Mark Immartino",
 		email: "mark.iammartino@gmail.com",
 		password: "fionashome301",
@@ -2307,25 +2390,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed83",
+		guest_id: "aswed86",
 		name: "Ryan & Lori",
 		email: "nowakrya@yahoo.com",
 		password: "fionashome301",
@@ -2335,25 +2418,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed84",
+		guest_id: "aswed87",
 		name: "David Song",
 		email: "drdavidsong@gmail.com",
 		password: "fionashome301",
@@ -2363,25 +2446,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed85",
+		guest_id: "aswed88",
 		name: "Ramesh & Shoba",
 		email: "chittur.ramesh@gmail.com",
 		password: "fionashome301",
@@ -2391,25 +2474,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed86",
+		guest_id: "aswed89",
 		name: "Veena & Gani",
 		email: "veenanayak@hotmail.com",
 		password: "fionashome301",
@@ -2419,25 +2502,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed87",
+		guest_id: "aswed90",
 		name: "Leena & Anand",
 		email: "Balsekar@aol.com",
 		password: "fionashome301",
@@ -2447,25 +2530,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed88",
+		guest_id: "aswed91",
 		name: "Maliakka & Ram",
 		email: "Malishetty18@gmail.com",
 		password: "fionashome301",
@@ -2475,25 +2558,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed89",
+		guest_id: "aswed92",
 		name: "Gayathri & Ram",
 		email: "Gailkamath@yahoo.com",
 		password: "fionashome301",
@@ -2503,25 +2586,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed90",
+		guest_id: "aswed93",
 		name: "Jay & Guru",
 		email: "gururao@yahoo.com",
 		password: "fionashome301",
@@ -2531,25 +2614,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed91",
+		guest_id: "aswed94",
 		name: "Shylaja & Lakoo",
 		email: "sliyer99@gmail.com",
 		password: "fionashome301",
@@ -2559,25 +2642,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed92",
+		guest_id: "aswed95",
 		name: "Vipuj & Amit",
 		email: "vipuj.shah@gmail.com",
 		password: "fionashome301",
@@ -2587,25 +2670,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed93",
+		guest_id: "aswed96",
 		name: "Anuthama & Raghu",
 		email: "anuttamaramani@gmail.com",
 		password: "fionashome301",
@@ -2615,25 +2698,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed94",
+		guest_id: "aswed97",
 		name: "Malathi",
 		email: "malathi.kashyap@yahoo.com",
 		password: "fionashome301",
@@ -2643,25 +2726,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed95",
+		guest_id: "aswed98",
 		name: "Shukla & Shirish",
 		email: "Shuklak@hotmail.com",
 		password: "fionashome301",
@@ -2671,25 +2754,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed96",
+		guest_id: "aswed99",
 		name: "Mona & Rajesh Char",
 		email: "monachar@hotmail.com",
 		password: "fionashome301",
@@ -2699,25 +2782,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed97",
+		guest_id: "aswed100",
 		name: "Tina & Ajay Madhavan",
 		email: "quantumpot@gmail.com",
 		password: "fionashome301",
@@ -2727,25 +2810,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed98",
+		guest_id: "aswed101",
 		name: "Usha & Ajay Balan",
 		email: "anbalan@yahoo.com",
 		password: "fionashome301",
@@ -2755,25 +2838,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed99",
+		guest_id: "aswed102",
 		name: "Geetha & Suresh Ramachandran",
 		email: "Surgee@yahoo.com",
 		password: "fionashome301",
@@ -2783,25 +2866,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed100",
+		guest_id: "aswed103",
 		name: "Tina & Sridhar Rao",
 		email: "mailspr@gmail.com",
 		password: "fionashome301",
@@ -2811,25 +2894,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed101",
+		guest_id: "aswed104",
 		name: "Sumi & Srikanth Kurkal ",
 		email: "Srikanthkurkal@gmail.com",
 		password: "fionashome301",
@@ -2839,25 +2922,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed102",
+		guest_id: "aswed105",
 		name: "Shanti & Seshadri",
 		email: "sesh1510@gmail.com",
 		password: "fionashome301",
@@ -2867,25 +2950,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed103",
+		guest_id: "aswed106",
 		name: "Anita & Kedar Navkal",
 		email: "anitanavkal@yahoo.com",
 		password: "fionashome301",
@@ -2895,25 +2978,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed104",
+		guest_id: "aswed107",
 		name: "Sonali & Manoj Thatte",
 		email: "sonalimthatte@yahoo.com",
 		password: "fionashome301",
@@ -2923,25 +3006,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed105",
+		guest_id: "aswed108",
 		name: "Ritu & Sanjeev",
 		email: "sanjivmahajan@yahoo.com",
 		password: "fionashome301",
@@ -2951,25 +3034,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed106",
+		guest_id: "aswed109",
 		name: "Geetu & Sanjay",
 		email: "Geetudalwani@yahoo.com",
 		password: "fionashome301",
@@ -2979,25 +3062,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed107",
+		guest_id: "aswed110",
 		name: "Rakhi & Ravi Gautam",
 		email: "rakhi.gautham@gmail.com",
 		password: "fionashome301",
@@ -3007,25 +3090,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed108",
+		guest_id: "aswed111",
 		name: "Pooja Soni",
 		email: "poojasoni1234@yahoo.com",
 		password: "fionashome301",
@@ -3035,25 +3118,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed109",
+		guest_id: "aswed112",
 		name: "Siva & Kala",
 		email: "kala@edifytech.usa",
 		password: "fionashome301",
@@ -3063,25 +3146,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed110",
+		guest_id: "aswed113",
 		name: "Beena",
 		email: "beena@edifytech.usa",
 		password: "fionashome301",
@@ -3091,25 +3174,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed111",
+		guest_id: "aswed114",
 		name: "Prince",
 		email: "prince@edifytech.usa",
 		password: "fionashome301",
@@ -3119,25 +3202,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed112",
+		guest_id: "aswed115",
 		name: "Sunil & Varsha",
 		email: "sunilqq@yahoo.com",
 		password: "fionashome301",
@@ -3147,25 +3230,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed113",
+		guest_id: "aswed116",
 		name: "Anju Jain & Ganesh Iyer",
 		email: "anju_jain@hotmail.com",
 		password: "fionashome301",
@@ -3175,25 +3258,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed114",
+		guest_id: "aswed117",
 		name: "Shreya & Mike Friend",
 		email: "shreyaukil@gmail.com",
 		password: "fionashome301",
@@ -3203,25 +3286,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed115",
+		guest_id: "aswed118",
 		name: "Anita Singh",
 		email: "anisingh@hotmail.com",
 		password: "fionashome301",
@@ -3231,25 +3314,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed116",
+		guest_id: "aswed119",
 		name: "Mohan Krishnan & Priya",
 		email: "mkrishnan@earthlink.net",
 		password: "fionashome301",
@@ -3259,25 +3342,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed117",
+		guest_id: "aswed120",
 		name: "Anupa & Amit",
 		email: "anupa1209@gmail.com",
 		password: "fionashome301",
@@ -3287,25 +3370,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed118",
+		guest_id: "aswed121",
 		name: "Sundar Nagarajan",
 		email: "sun.nagarajan@gmail.com",
 		password: "fionashome301",
@@ -3315,25 +3398,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed119",
+		guest_id: "aswed122",
 		name: "Krish Nagarajan",
 		email: "krishnan.nagarajan@gmail.com",
 		password: "fionashome301",
@@ -3343,25 +3426,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed120",
+		guest_id: "aswed123",
 		name: "Ganesh Pai",
 		email: "ganesh@suvigaventures.com",
 		password: "fionashome301",
@@ -3371,25 +3454,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed121",
+		guest_id: "aswed124",
 		name: "Sanjeev & Poonam",
 		email: "sanjeev.harvest@gmail.com",
 		password: "fionashome301",
@@ -3399,25 +3482,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed122",
+		guest_id: "aswed125",
 		name: "Sunish & Dimpy",
 		email: "sunish@gmail.com",
 		password: "fionashome301",
@@ -3427,25 +3510,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed123",
+		guest_id: "aswed126",
 		name: "Vikas & Payal",
 		email: "tovikasdhingra@gmail.com",
 		password: "fionashome301",
@@ -3455,25 +3538,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed124",
+		guest_id: "aswed127",
 		name: "Srivats & Kritika",
 		email: "cricmaniac@gmail.com",
 		password: "fionashome301",
@@ -3483,25 +3566,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed125",
+		guest_id: "aswed128",
 		name: "Ganesh Pai",
 		email: "ganeshragpai@outlook.com",
 		password: "fionashome301",
@@ -3511,25 +3594,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed126",
+		guest_id: "aswed129",
 		name: "Annick",
 		email: "annick.reis@gmail.com",
 		password: "fionashome301",
@@ -3539,25 +3622,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed127",
+		guest_id: "aswed130",
 		name: "Cathy & Chris",
 		email: "ccjeriehs@gmail.com",
 		password: "fionashome301",
@@ -3567,25 +3650,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed128",
+		guest_id: "aswed131",
 		name: "Rya & Arman Yamboa",
 		email: "ryambao123@gmail.com",
 		password: "fionashome301",
@@ -3595,25 +3678,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed129",
+		guest_id: "aswed132",
 		name: "Paul Soderland",
 		email: "paul_soderlind@yahoo.com",
 		password: "fionashome301",
@@ -3623,25 +3706,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed130",
+		guest_id: "aswed133",
 		name: "Renee & Mike",
 		email: "reneelarmon@yahoo.com",
 		password: "fionashome301",
@@ -3651,25 +3734,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed131",
+		guest_id: "aswed134",
 		name: "Karen & Paul",
 		email: "kmhrich@comcast.net",
 		password: "fionashome301",
@@ -3679,25 +3762,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed132",
+		guest_id: "aswed135",
 		name: "Rob & Collette",
 		email: "mahanesr@gmail.com",
 		password: "fionashome301",
@@ -3707,25 +3790,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed133",
+		guest_id: "aswed136",
 		name: "Sheila & Joe",
 		email: "sawilfinger@gmail.com",
 		password: "fionashome301",
@@ -3735,81 +3818,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed134",
-		name: "Greg H & Cindy",
-		email: "hagermang@aol.com",
-		password: "fionashome301",
-		ReceptionOnly: true,
-		SangeetReceptionOnly: false,
-		SanMuhRec: false,
-		AllEvents: false,
-		AllUSJapan: false,
-		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
-		isAttendingHaldi: "no",
-		haldiPartySize: 0,
-		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
-		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
-		isAttendingReception: "no",
-		receptionPartySize: 0,
-		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
-		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
-		lastRSVPUpdateTime: "2024-05-26 10:00",
-		partyMembers: "avani",
-		allergies: "none"
-	},
-	{
-		guest_id: "aswed135",
-		name: "Phaly & Arkadius",
-		email: "phaly.pichota@gmail.com",
-		password: "fionashome301",
-		ReceptionOnly: true,
-		SangeetReceptionOnly: false,
-		SanMuhRec: false,
-		AllEvents: false,
-		AllUSJapan: false,
-		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
-		isAttendingHaldi: "no",
-		haldiPartySize: 0,
-		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
-		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
-		isAttendingReception: "no",
-		receptionPartySize: 0,
-		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
-		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
-		lastRSVPUpdateTime: "2024-05-26 10:00",
-		partyMembers: "avani",
-		allergies: "none"
-	},
-	{
-		guest_id: "aswed136",
+		guest_id: "aswed137",
 		name: "Gopi & Shreesha",
 		email: "raj@valuemomentum.com",
 		password: "fionashome301",
@@ -3819,25 +3846,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed137",
+		guest_id: "aswed138",
 		name: "KK ",
 		email: "kalyan.kodali@valuemomentum.com",
 		password: "fionashome301",
@@ -3847,25 +3874,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed138",
+		guest_id: "aswed139",
 		name: "Venkat",
 		email: "venkat.pydimarri@valuemomentum.com",
 		password: "fionashome301",
@@ -3875,25 +3902,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed139",
+		guest_id: "aswed140",
 		name: "GK",
 		email: "gopi.gade@valuemomentum.com",
 		password: "fionashome301",
@@ -3903,25 +3930,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed140",
+		guest_id: "aswed141",
 		name: "Swarup Ghosh",
 		email: "swarup.ghosh@valuemomentum.com",
 		password: "fionashome301",
@@ -3931,25 +3958,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed141",
+		guest_id: "aswed142",
 		name: "Vijay M",
 		email: "vijay.muthiah@outlook.com",
 		password: "fionashome301",
@@ -3959,25 +3986,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed142",
+		guest_id: "aswed143",
 		name: "Suzy Cera",
 		email: "suzycerra@gmail.com",
 		password: "fionashome301",
@@ -3987,25 +4014,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed143",
+		guest_id: "aswed144",
 		name: "Patty",
 		email: "pattysaccone@comcast.net",
 		password: "fionashome301",
@@ -4015,25 +4042,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed144",
+		guest_id: "aswed145",
 		name: "Elena",
 		email: "ekdrbg0103@gmail.com",
 		password: "fionashome301",
@@ -4043,25 +4070,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed145",
+		guest_id: "aswed146",
 		name: "Ellen",
 		email: "meuch@msn.com",
 		password: "fionashome301",
@@ -4071,25 +4098,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed146",
+		guest_id: "aswed147",
 		name: "Sarah & Kurt",
 		email: "tri74girl@gmail.com",
 		password: "fionashome301",
@@ -4099,25 +4126,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed147",
+		guest_id: "aswed148",
 		name: "Erin",
 		email: "riemcald@yahoo.com",
 		password: "fionashome301",
@@ -4127,25 +4154,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed148",
+		guest_id: "aswed149",
 		name: "Susan",
 		email: "hamiltonsec@gmail.com",
 		password: "fionashome301",
@@ -4155,25 +4182,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed149",
+		guest_id: "aswed150",
 		name: "Manoj (Citi)",
 		email: "manoj.enas@citi.com",
 		password: "fionashome301",
@@ -4183,25 +4210,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed150",
+		guest_id: "aswed151",
 		name: "Dan (Citi)",
 		email: "daniel.demassi@citi.com",
 		password: "fionashome301",
@@ -4211,25 +4238,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed151",
+		guest_id: "aswed152",
 		name: "William Lu (Citi)",
 		email: "zhenlin.lu@citi.com",
 		password: "fionashome301",
@@ -4239,25 +4266,25 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
 	},
 	{
-		guest_id: "aswed152",
+		guest_id: "aswed153",
 		name: "Vivek Chopra",
 		email: "vivekchopra12@gmail.com",
 		password: "fionashome301",
@@ -4267,19 +4294,19 @@ const guest_table =
 		AllEvents: false,
 		AllUSJapan: false,
 		isAttendingMehendi: "no",
-		mehendiPartySize: 0,
+		mehendiPartySize: "0",
 		isAttendingHaldi: "no",
-		haldiPartySize: 0,
+		haldiPartySize: "0",
 		isAttendingSangeet: "no",
-		sangeetPartySize: 0,
+		sangeetPartySize: "0",
 		isAttendingMuhurtham: "no",
-		muhurthamPartySize: 0,
+		muhurthamPartySize: "0",
 		isAttendingReception: "no",
-		receptionPartySize: 0,
+		receptionPartySize: "0",
 		isAttendingShinzenshiki: "no",
-		shinzenshikiPartySize: 0,
+		shinzenshikiPartySize: "0",
 		isAttendingHiroen: "no",
-		hiroenPartySize: 0,
+		hiroenPartySize: "0",
 		lastRSVPUpdateTime: "2024-05-26 10:00",
 		partyMembers: "avani",
 		allergies: "none"
