+++ app/wedding/rsvp/page.tsx
@@ -12,7 +12,7 @@ export default async function Page() {
         <main>
             <div className="bg-[url('/sakura_border.png')] bg-contain container flex flex-col justify-evenly p-4 mx-auto max-width">
                 <h1 className={` ${birthstonebounce.className} text-2xl md:text-7xl text-center md:mb-4`}>We hope you can join us!</h1>
-                <h1 className="text-md text-center"><br/>Kindly respond by July 6, 2024</h1>
+                <h1 className="text-md text-center"><br/>Kindly respond by July 15<sup>th</sup>, 2024</h1>
             </div>
             <RSVPForm invites={invitedEvents}/>
         </main>
