+++ package.json
@@ -6,7 +6,7 @@
     "prettier": "prettier --write --ignore-unknown .",
     "prettier:check": "prettier --check --ignore-unknown .",
     "start": "next start",
-    "seed": "node -r dotenv/config ./scripts/test_seed.js",
+    "seed": "node -r dotenv/config ./scripts/seed.js",
     "lint": "next lint"
   },
   "dependencies": {
