+++ scripts/seed.js
@@ -1,44 +1,117 @@
 const { db } = require('@vercel/postgres');
 const {
-  invoices,
-  customers,
-  revenue,
-  users,
-} = require('../app/lib/placeholder-data.js');
+  guest_table
+} = require('../app/lib/guests.js');
 const bcrypt = require('bcrypt');
 
 async function seedUsers(client) {
   try {
     await client.sql`CREATE EXTENSION IF NOT EXISTS "uuid-ossp"`;
-    // Create the "users" table if it doesn't exist
+
     const createTable = await client.sql`
-      CREATE TABLE IF NOT EXISTS users (
-        id UUID DEFAULT uuid_generate_v4() PRIMARY KEY,
-        name VARCHAR(255) NOT NULL,
-        email TEXT NOT NULL UNIQUE,
-        password TEXT NOT NULL
+      CREATE TABLE IF NOT EXISTS guests (
+      guest_id VARCHAR(225) NOT NULL PRIMARY KEY,
+      name VARCHAR(255) NOT NULL,
+      email TEXT NOT NULL UNIQUE,
+      password TEXT NOT NULL,
+	  ReceptionOnly BOOLEAN NOT NULL,
+	  SangeetReceptionOnly BOOLEAN NOT NULL,
+	  SanMuhRec BOOLEAN NOT NULL,
+	  AllEvents BOOLEAN NOT NULL,
+	  AllUSJapan BOOLEAN NOT NULL,
+	  isAttendingMehendi VARCHAR(225) NOT NULL,
+	  mehendiPartySize INT NOT NULL,
+	  isAttendingHaldi VARCHAR(225) NOT NULL,
+	  haldiPartySize INT NOT NULL,
+	  isAttendingSangeet VARCHAR(225) NOT NULL,
+	  sangeetPartySize INT NOT NULL,
+	  isAttendingMuhurtham VARCHAR(225) NOT NULL,
+	  muhurthamPartySize INT NOT NULL,
+	  isAttendingReception VARCHAR(225) NOT NULL,
+	  receptionPartySize INT NOT NULL,
+	  isAttendingShinzenshiki VARCHAR(225) NOT NULL,
+	  shinzenshikiPartySize INT NOT NULL,
+	  isAttendingHiroen VARCHAR(225) NOT NULL,
+	  hiroenPartySize INT NOT NULL,
+	  lastRSVPUpdateTime TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+	  partyMembers TEXT NOT NULL,
+	  allergies TEXT NOT NULL
       );
     `;
 
-    console.log(`Created "users" table`);
+    console.log(`Created "guests" table`);
 
-    // Insert data into the "users" table
-    const insertedUsers = await Promise.all(
-      users.map(async (user) => {
+    // Insert data into the "guests" table
+    const insertedGuests = await Promise.all(
+      guest_table.map(async (user) => {
         const hashedPassword = await bcrypt.hash(user.password, 10);
         return client.sql`
-        INSERT INTO users (id, name, email, password)
-        VALUES (${user.id}, ${user.name}, ${user.email}, ${hashedPassword})
-        ON CONFLICT (id) DO NOTHING;
+        INSERT INTO guests (
+			guest_id, 
+			name, 
+			email, 
+			password, 
+			ReceptionOnly,
+			SangeetReceptionOnly,
+			SanMuhRec,
+			AllEvents,
+			AllUSJapan,
+			isAttendingMehendi, 
+			mehendiPartySize, 
+			isAttendingHaldi, 
+			haldiPartySize, 
+			isAttendingSangeet, 
+			sangeetPartySize, 
+			isAttendingMuhurtham, 
+			muhurthamPartySize, 
+			isAttendingReception, 
+			receptionPartySize, 
+			isAttendingShinzenshiki,
+			shinzenshikiPartySize, 
+			isAttendingHiroen, 
+			hiroenPartySize, 
+			lastRSVPUpdateTime, 
+			partyMembers,
+			allergies 
+		)
+        VALUES (
+			${user.guest_id}, 
+			${user.name}, 
+			${user.email}, 
+			${hashedPassword}, 
+			${user.ReceptionOnly},
+			${user.SangeetReceptionOnly},
+			${user.SanMuhRec},
+			${user.AllEvents},
+			${user.AllUSJapan},
+			${user.isAttendingMehendi}, 
+			${user.mehendiPartySize},
+			${user.isAttendingHaldi}, 
+			${user.haldiPartySize}, 
+			${user.isAttendingSangeet}, 
+			${user.sangeetPartySize},
+			${user.isAttendingMuhurtham}, 
+			${user.muhurthamPartySize},
+			${user.isAttendingReception}, 
+			${user.receptionPartySize},
+			${user.isAttendingShinzenshiki}, 
+			${user.shinzenshikiPartySize},
+			${user.isAttendingHiroen}, 
+			${user.hiroenPartySize},
+			${user.lastRSVPUpdateTime}, 
+			${user.partyMembers},
+			${user.allergies} 
+		)
+        ON CONFLICT (guest_id) DO NOTHING;
       `;
       }),
     );
 
-    console.log(`Seeded ${insertedUsers.length} users`);
+    console.log(`Seeded ${insertedGuests.length} guests`);
 
     return {
       createTable,
-      users: insertedUsers,
+      users: insertedGuests,
     };
   } catch (error) {
     console.error('Error seeding users:', error);
@@ -46,127 +119,21 @@ async function seedUsers(client) {
   }
 }
 
-async function seedInvoices(client) {
+async function removeTable(client){
   try {
-    await client.sql`CREATE EXTENSION IF NOT EXISTS "uuid-ossp"`;
+    await client.sql`DROP TABLE IF EXISTS {table_name};`;
 
-    // Create the "invoices" table if it doesn't exist
-    const createTable = await client.sql`
-    CREATE TABLE IF NOT EXISTS invoices (
-    id UUID DEFAULT uuid_generate_v4() PRIMARY KEY,
-    customer_id UUID NOT NULL,
-    amount INT NOT NULL,
-    status VARCHAR(255) NOT NULL,
-    date DATE NOT NULL
-  );
-`;
-
-    console.log(`Created "invoices" table`);
-
-    // Insert data into the "invoices" table
-    const insertedInvoices = await Promise.all(
-      invoices.map(
-        (invoice) => client.sql`
-        INSERT INTO invoices (customer_id, amount, status, date)
-        VALUES (${invoice.customer_id}, ${invoice.amount}, ${invoice.status}, ${invoice.date})
-        ON CONFLICT (id) DO NOTHING;
-      `,
-      ),
-    );
-
-    console.log(`Seeded ${insertedInvoices.length} invoices`);
-
-    return {
-      createTable,
-      invoices: insertedInvoices,
-    };
   } catch (error) {
-    console.error('Error seeding invoices:', error);
+    console.error('Error deleteing provided table. ', error);
     throw error;
   }
+
 }
-
-async function seedCustomers(client) {
-  try {
-    await client.sql`CREATE EXTENSION IF NOT EXISTS "uuid-ossp"`;
-
-    // Create the "customers" table if it doesn't exist
-    const createTable = await client.sql`
-      CREATE TABLE IF NOT EXISTS customers (
-        id UUID DEFAULT uuid_generate_v4() PRIMARY KEY,
-        name VARCHAR(255) NOT NULL,
-        email VARCHAR(255) NOT NULL,
-        image_url VARCHAR(255) NOT NULL
-      );
-    `;
-
-    console.log(`Created "customers" table`);
-
-    // Insert data into the "customers" table
-    const insertedCustomers = await Promise.all(
-      customers.map(
-        (customer) => client.sql`
-        INSERT INTO customers (id, name, email, image_url)
-        VALUES (${customer.id}, ${customer.name}, ${customer.email}, ${customer.image_url})
-        ON CONFLICT (id) DO NOTHING;
-      `,
-      ),
-    );
-
-    console.log(`Seeded ${insertedCustomers.length} customers`);
-
-    return {
-      createTable,
-      customers: insertedCustomers,
-    };
-  } catch (error) {
-    console.error('Error seeding customers:', error);
-    throw error;
-  }
-}
-
-async function seedRevenue(client) {
-  try {
-    // Create the "revenue" table if it doesn't exist
-    const createTable = await client.sql`
-      CREATE TABLE IF NOT EXISTS revenue (
-        month VARCHAR(4) NOT NULL UNIQUE,
-        revenue INT NOT NULL
-      );
-    `;
-
-    console.log(`Created "revenue" table`);
-
-    // Insert data into the "revenue" table
-    const insertedRevenue = await Promise.all(
-      revenue.map(
-        (rev) => client.sql`
-        INSERT INTO revenue (month, revenue)
-        VALUES (${rev.month}, ${rev.revenue})
-        ON CONFLICT (month) DO NOTHING;
-      `,
-      ),
-    );
-
-    console.log(`Seeded ${insertedRevenue.length} revenue`);
-
-    return {
-      createTable,
-      revenue: insertedRevenue,
-    };
-  } catch (error) {
-    console.error('Error seeding revenue:', error);
-    throw error;
-  }
-}
-
 async function main() {
   const client = await db.connect();
 
+  //await removeTable(client);
   await seedUsers(client);
-  await seedCustomers(client);
-  await seedInvoices(client);
-  await seedRevenue(client);
 
   await client.end();
 }
