+++ tsconfig.json
@@ -28,7 +28,7 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     "app/lib/placeholder-data.js",
-    "scripts/seed.js"
+    "scripts/tut_seed.js"
   ],
   "exclude": ["node_modules"]
 }
